requirements.txt
networkx==2.8.4
numpy==1.23.0
pygame==2.1.2
PyOpenGL==3.1.6
